requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
Django==6.0.4
idna==3.12
pillow==12.2.0
pyTelegramBotAPI==4.33.0
python-dotenv==1.2.2
requests==2.33.1
sqlparse==0.5.5
tzdata==2026.1
urllib3==2.6.3
